Fixed compile warnings (most were due to implicit casts).
Updated camera to newer simpler pythagoras camera.

Updated the dependencies, and included them in the project file to ease initial compilation.

+++ sdl_exp/Axis.cpp
@@ -3,7 +3,7 @@
 #include <stdio.h>
 #include "SDL/SDL_opengl.h"
 
-Axis::Axis(double length){
+Axis::Axis(float length){
 	this->length = length;
 }
